+++ plugins/voxelman/worldinteraction/plugin.d
@@ -19,7 +19,7 @@ import voxelman.core.events;
 import voxelman.core.packets;
 import voxelman.world.storage.coordinates;
 import voxelman.world.storage.volume;
-import voxelman.world.storage.blockentityaccess;
+import voxelman.blockentity.blockentityaccess;
 
 import voxelman.block.plugin;
 import voxelman.eventdispatcher.plugin;